+++ src/Five.java
@@ -1,7 +1,7 @@
 public class Five {
-    private int five = 5;
+    private short five = 5;
 
-    public int fiveCoins(int change) {
-        return change / five;
+    public short fiveCoins(short change) {
+        return (short)(change / five);
     }
 }
