+++ src/Ten.java
@@ -1,7 +1,7 @@
 public class Ten {
-    private int ten = 10;
+    private short ten = 10;
 
-    public int tenCoins(int change) {
-        return change / ten;
+    public short tenCoins(short change) {
+        return (short)(change / ten);
     }
 }
